# api/openapi/v1/keyip.yaml
openapi: "3.0.3"
info:
  title: KeyIP-Intelligence API
  version: "1.0.0"
  description: |
    KeyIP-Intelligence is an AI-driven intellectual property lifecycle management platform
    designed for the OLED organic materials industry. It integrates molecular-level chemical
    understanding with patent law intelligence to support patent mining, infringement
    monitoring, portfolio optimization, and lifecycle management.

    ## Authentication

    All API endpoints require Bearer token authentication unless otherwise specified.
    Tokens are issued via the configured identity provider (Keycloak / OIDC-compatible).

    ## Pagination

    List endpoints support page-based pagination via `page` and `page_size` query parameters.
    Responses include pagination metadata.

    ## Error Handling

    All errors follow a unified structure:
    ```json
    {
      "code": "NOT_FOUND",
      "message": "The specified resource does not exist"
    }
    ```
  contact:
    name: KeyIP-Intelligence Open Source Community
    url: https://github.com/turtacn/KeyIP-Intelligence
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.keyip.example.com/api/v1
    description: Production server
  - url: https://staging-api.keyip.example.com/api/v1
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/workspaces:
    get:
      tags: [Collaboration]
      summary: List workspaces
      operationId: listWorkspaces
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Collaboration]
      summary: Create workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/workspaces/{id}:
    get:
      tags: [Collaboration]
      summary: Get workspace
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Collaboration]
      summary: Update workspace
      operationId: updateWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Collaboration]
      summary: Delete workspace
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{id}/documents:
    get:
      tags: [Collaboration]
      summary: List shared documents
      operationId: listSharedDocuments
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Shared document list
          content:
            application/json:
              schema:
                type: object

    post:
      tags: [Collaboration]
      summary: Share document
      operationId: shareDocument
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareDocumentRequest"
      responses:
        "201":
          description: Document shared
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{id}/members:
    get:
      tags: [Collaboration]
      summary: List workspace members
      operationId: listWorkspaceMembers
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Collaboration]
      summary: Invite member
      operationId: inviteMember
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{id}/members/{memberId}:
    delete:
      tags: [Collaboration]
      summary: Remove workspace member
      operationId: removeWorkspaceMember
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{id}/members/{memberId}/role:
    put:
      tags: [Collaboration]
      summary: Update member role
      operationId: updateMemberRole
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{id}/shared-resource:
    get:
      tags: [Collaboration]
      summary: Get shared resource
      operationId: getSharedResource
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Shared resource
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{id}/shares/{shareId}:
    delete:
      tags: [Collaboration]
      summary: Revoke share link
      operationId: revokeShareLink
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: shareId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Share link revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Reporting
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: Page number (starts at 1).
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      description: Number of items per page (default 20, max 100).
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    WorkspaceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication token missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.

    # -------------------------------------------------------------------------
    # Health
    # -------------------------------------------------------------------------
...
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [private, internal, partner]
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [private, internal, partner]
    WorkspaceResponse:
      type: object
      description: Workspace detail. Shape determined by the service layer.
    WorkspaceListResponse:
      type: object
      description: Paginated workspace list. Shape determined by the service layer.
    ShareDocumentRequest:
      type: object
      required: [document_id]
      properties:
        document_id:
          type: string
        workspace_id:
          type: string
        enable_watermark:
          type: boolean
          default: false
        max_downloads:
          type: integer
        expires_in_hours:
          type: integer
    InviteMemberRequest:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [owner, admin, editor, viewer]
    UpdateMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, admin, editor, viewer]

    # -------------------------------------------------------------------------
    # Reporting
    # -------------------------------------------------------------------------
...